+++ pages/vue/file-picker/file-picker.vue
@@ -10,11 +10,17 @@
 		</uni-section>
 		<uni-section title="只选择视频" type="line">
 			<view class="example-body">
-				<uni-file-picker limit="9" file-mediatype="video" title="最多选择9个视频" :sourceType="sourceType"></uni-file-picker>
+				<uni-file-picker limit="9" file-mediatype="video" title="最多选择9个视频" :source-type="sourceType"></uni-file-picker>
 			</view>
 		</uni-section>
-
-
+		<uni-section title="自定义图片和视频选择的来源" type="line">
+			<view class="example-body">
+				<uni-file-picker limit="9" title="从相册选图" :source-type="['album']"></uni-file-picker>
+			</view>
+			<view class="example-body">
+				<uni-file-picker limit="9" title="使用相机" file-mediatype="video" :source-type="['camera']"></uni-file-picker>
+			</view>
+		</uni-section>
 		<!-- #ifdef H5 || MP-WEIXIN -->
 		<uni-section title="选择任意文件" type="line">
 			<view class="example-body">
@@ -48,7 +54,6 @@
 		data() {
 			return {
 				sourceType:['camera','album'],
-
 				imageStyles: {
 					width: 64,
 					height: 64,